Move penalty-type CRUD and penalty-history routes to dedicated blueprint
Extracts nine routes from _api.py into a new app/routes/penalty_types.py
blueprint at /_api: five tournament-scoped routes (penalty-types CRUD
plus per-player penalty-history) and four league-scoped routes (penalty-
types CRUD). Tournament and league penalty types are parallel concepts,
so co-locating them in one file mirrors the existing grouping rule.

_check_to and _require_league are imported transitionally from _api.py
and will be replaced when later changes move those helpers to permanent
homes. URL surface is unchanged.
